Fix mobile firefly renderer by reimplementing it on top of particles. It's likely this entity will go away entirely when we get 1.16's support for biome-ambience particles

+++ src/main/java/twilightforest/client/particle/ParticleFirefly.java
@@ -45,7 +45,7 @@ public class ParticleFirefly extends SpriteTexturedParticle {
 	}
 
 	public float getGlowBrightness() {
-		float lifeTime = maxAge - age;
+		int lifeTime = maxAge - age;
 		if (lifeTime <= halfLife) {
 			return (float) lifeTime / (float) halfLife;
 		} else {
